+++ libs/designsystem/src/lib/components/charts/charts.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { ChartComponent } from './chart';
-import { ChartConfigService } from './shared/base-chart';
+import { ChartConfigService } from './shared';
 import { StockChartComponent } from './stock-chart';
 
 @NgModule({
