+++ libs/designsystem/src/lib/components/charts/shared/base-chart/base-chart.component.ts
@@ -13,8 +13,7 @@ import { AnnotationOptions } from 'chartjs-plugin-annotation';
 
 import { ChartDataset, ChartHighlightedElements, ChartLabel, ChartType } from '../';
 import { ResizeObserverFactory } from '../../../shared';
-
-import { ChartJSService } from './chart-js/chart-js.service';
+import { ChartJSService } from '../chart-js-service/';
 
 @Component({
   selector: 'kirby-base-chart',
