+++ libs/designsystem/src/lib/components/charts/shared/base-chart/index.ts
@@ -1,3 +1,1 @@
 export { BaseChartComponent } from './base-chart.component';
-export { ChartJSService } from './chart-js/chart-js.service';
-export { ChartConfigService } from './configs/chart-config.service';
