+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/configs/annotations.config.ts
