+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/configs/global-defaults.config.ts
