+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/configs/interaction-functions-extensions.config.ts
