+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/configs/type.config.ts
