+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/index.ts
@@ -1,0 +1,3 @@
+export { ChartJSService } from './chart-js.service';
+export { ChartConfigService } from './configs/chart-config.service';
+export * from './test-utils';
